update LanguageColor (#56)
* update LanguageColor

* update pods
